train/requirements.txt
chainer<6.0.0
cupy<6.0.0
onnx-chainer
pillow-simd
